// web/src/api/model/libraryWork.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Scholia API
 * OpenAPI spec version: 0.1.0
 */
import type { LibraryVersion } from "./libraryVersion";

export interface LibraryWork {
    co_authors: string[];
    /** @nullable */
    editor_names?: string[] | null;
    /** @nullable */
    publication_year?: number | null;
    title: string;
    versions: LibraryVersion[];
    work_id: string;
}